+++ packages/framework/tests/Bootstrap/cache/packages.php
@@ -1,8 +1,0 @@
-<?php
-
-return array (
-  'fingerprint' => '3ebfadea87a22033ee37699ff062f1018464cc7de3b5b7079967dc76b5400463',
-  'providers' => 
-  array (
-  ),
-);